refactor remote media types
