+++ src/Kunstmaan/MediaBundle/Form/RemoteVideo/RemoteVideoType.php
@@ -2,11 +2,8 @@
 
 namespace Kunstmaan\MediaBundle\Form\RemoteVideo;
 
-use Kunstmaan\MediaBundle\Repository\FolderRepository;
-use Symfony\Component\Form\AbstractType;
+use Kunstmaan\MediaBundle\Form\AbstractRemoteType;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\Form\FormEvent;
-use Symfony\Component\Form\FormEvents;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 use Symfony\Component\Validator\Constraints\NotBlank;
@@ -14,7 +11,7 @@ use Symfony\Component\Validator\Constraints\NotBlank;
 /**
  * RemoteVideoType
  */
-class RemoteVideoType extends AbstractType
+class RemoteVideoType extends AbstractRemoteType
 {
 
     /**
@@ -24,6 +21,7 @@ class RemoteVideoType extends AbstractType
 
     /**
      * Constructor, gets the RemoteVideo configuration
+     *
      * @param array $configuration
      */
     public function __construct($configuration = array())
@@ -38,11 +36,27 @@ class RemoteVideoType extends AbstractType
      * top most type. Type extensions can further modify the form.
      *
      * @param FormBuilderInterface $builder The form builder
-     * @param array $options The options
+     * @param array                $options The options
      *
      * @see FormTypeExtensionInterface::buildForm()
      */
     public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        parent::buildForm($builder, $options);
+
+        $builder
+            ->add(
+                'type',
+                'choice',
+                array(
+                    'choices'     => $this->getRemoteVideoChoices(),
+                    'constraints' => array(new NotBlank()),
+                    'required'    => true
+                )
+            );
+    }
+
+    protected function getRemoteVideoChoices()
     {
         $choices = array();
         if (count($this->configuration)) {
@@ -54,72 +68,7 @@ class RemoteVideoType extends AbstractType
             }
         }
 
-        $builder
-            ->add(
-                'name',
-                'text',
-                array(
-                    'constraints' => array(new NotBlank()),
-                    'required' => true
-                )
-            )
-            ->add(
-                'code',
-                'text',
-                array(
-                    'constraints' => array(new NotBlank()),
-                    'required' => true
-                )
-            )
-            ->add(
-                'type',
-                'choice',
-                array(
-                    'choices' => $choices,
-                    'constraints' => array(new NotBlank()),
-                    'required' => true
-                )
-            )
-            ->add(
-                'copyright',
-                'text',
-                array(
-                    'required' => false
-                )
-            )
-            ->add(
-                'description',
-                'textarea',
-                array(
-                    'required' => false
-                )
-            );
-
-        $builder->addEventListener(
-            FormEvents::PRE_SET_DATA,
-            function (FormEvent $event) {
-                $helper = $event->getData();
-                $form = $event->getForm();
-
-                // Make sure file field is when creating new (not persisted) objects
-                if (null !== $helper->getMedia()->getId()) {
-                    // Allow changing folder on edit
-                    $form->add(
-                        'folder',
-                        'entity',
-                        array(
-                            'class' => 'KunstmaanMediaBundle:Folder',
-                            'property' => 'optionLabel',
-                            'query_builder' => function (FolderRepository $er) {
-                                return $er->selectFolderQueryBuilder()
-                                    ->andWhere('f.parent IS NOT NULL');
-                            },
-                            'required' => true,
-                        )
-                    );
-                }
-            }
-        );
+        return $choices;
     }
 
     /**
